Move IQueryColumn to its own file
Gets rid of circular dependency between QueryColumn and ViewInfo.
